Implement downloading of package tarballs
When working with GitHub repositories instead of cloning them now by
default will be downloaded tarball using GitHub REST API. It is provided
a fall-back option `-t, --no-tarballs` for disabling this behavior.

Additionally, a bug in calculating package checksums is fixed. The list
of package files now is being sorted before calculation of the checksum
to ensure common behavior with different versions of the VCS tools and
independence from the download method.

Related to nim-lang/nimble#127