Add `develop --global` option
An option for creating the old-style Nimble link files is added. Those
files are created by `nimble develop` when called with the `--global`
option. They are placed in a special `links` directory and they are read
by Nim in order to have some kind of develop mode functionality when
compiling stand-alone files which are not part of a Nimble package.
Nimble itself ignores this new directory and files in it. When building
with Nimble the new local develop mode should be used instead.

Related to nim-lang/nimble#948